moved/rename options data
